Init 1.10 Update
+Rewrite of the Core -> Switched from a tag based system to a BlockID system.
+Changed the way you would add Custom Tags/Blocks
-Removed Tags
+Splitted the Config into 2 seperate files which are located in /YOU MC FOLDER/config/Qubik Studios Mods/SneakyTreeGrowing/[main.toml | blockList.toml]
+ Readded Tags as a way to filter the targets
A array for the tags can now be found in the blockList.toml config file.
+ Added the ability that the mod consumes Bone meal. Can be changed in the Config (Default = false)
Bone meal will be consumed from the player inventory. As its default value it is ment to be used in a Modpack to nerf the usage of STG.
Player will recieve a message when no more bone meal is in the inventory and effect will be cancelled.
This feature setting can be found in the main.toml
+ Added the feature of hunger loss. Can be changed in the Config (Default = true)
The player will get a random amount of hunger removed between 0 and 2 levels of food.
Effect won't be applied the whole time! You have a 75% change that it will work.
This feature setting can be found in the main.toml
+ Added a Dev Build Message when player joins a world
(Will be removed in public builds)
++ Changed the chance math formular to reduce the effect.
Now the chance is [x < chance / 120] instead of [x < chance / 100] to make the effect come less often.
x is a random generated Number.
- Removed the Version string in-code. No usage.